% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_predictors.R
\name{adj_angle_sum_variance}
\alias{adj_angle_sum_variance}
\title{Variance of the sum of adjacent angles}
\usage{
adj_angle_sum_variance(shape)
}
\arguments{
\item{shape}{- the shape for which to calculate the predictor}
}
\value{
The variance of the sum of adjacent angles
}
\description{
In paralelograms, the sum of all adjacent angle pairs is 180deg
}
