% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_predictors.R
\name{angle_variance}
\alias{angle_variance}
\title{Variance of the angles at vertices}
\usage{
angle_variance(shape)
}
\arguments{
\item{shape}{- the shape for which to calculate the predictor}
}
\value{
The variance of the angles at vertices
}
\description{
The idea is that e.g. rectangles have same angles
}
