% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_classifier.R
\name{classify_shape}
\alias{classify_shape}
\title{Classify a shape}
\usage{
classify_shape(shape)
}
\arguments{
\item{shape}{- the shape to classify}
}
\value{
A character vector with the code and long name of the shape class
}
\description{
Classifies the shape into one of the categories which can be seen from
shape_code_map, or Unknown if it has more than one contour or > 6 sides
}
