% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mines.R
\name{clip}
\alias{clip}
\title{Clip image border}
\usage{
clip(im, amount = 0)
}
\arguments{
\item{im}{- the image to clip}

\item{amount}{- the number of pixels to clip form each side, as numeric, or,
if character, the percentage of the smaller dimenision of
the image (e.g. clip(im, "5%"))}
}
\value{
The clipped image
}
\description{
Clip image border
}
