% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mines.R
\name{extract_fields}
\alias{extract_fields}
\title{Extract fields from an image, given the boundaries}
\usage{
extract_fields(im, boundaries = get_boundaries(im))
}
\arguments{
\item{im}{- the image from which to extract fields, using the boundaries}

\item{boundaries}{- the object of class 'bounds' (from get_boundaries
function), containing x and y ranges of field coordinates}
}
\value{
A list containing containing 'field' objects, i.e. lists with
        matrix coordinates of the field and the image of the field
}
\description{
Extract fields from an image, given the boundaries
}
