% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mines.R
\name{get_boundaries}
\alias{get_boundaries}
\title{Get coordinates of boundaries of fields}
\usage{
get_boundaries(im, eps = 0.1, prob = 0.8)
}
\arguments{
\item{im}{- the image from which to determine boundaries (must have exactly
one color channel)}

\item{eps}{- the highest value to treat as black}

\item{prob}{- the ratio of number of 'black' pixels in a line to the length
of the line}
}
\value{
An object (list) of class 'bounds' containing 81 elements, each
        having a $x and $y attributes giving, respectively, the x and y
        range of the boundaries.
}
\description{
Determines the x and y ranges of each field
}
