% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mines.R
\name{get_predictors}
\alias{get_predictors}
\title{Calculate specified predictors for given fields}
\usage{
get_predictors(predictors, fields)
}
\arguments{
\item{predictors}{- a character vector containing the names of functions that
take a field$image and return the value of a predictor}

\item{fields}{- a list containing objects of class 'field', usually got from
the extract_fields function}
}
\value{
A matrix with each row being the calculated predictors for a field,
        populated column-by-column, just as a regular R matrix, so the
        (i,j) field predictors are in the (i + 9*(j-1))-th row of the matrix.
}
\description{
Takes a vector of functions which represent predictors and applies them to
the given list of fields
}
