% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_predictors.R
\name{get_shape_predictors}
\alias{get_shape_predictors}
\title{Get predictors for a shape}
\usage{
get_shape_predictors(shape, df = FALSE, na_replacement = NA)
}
\arguments{
\item{shape}{- the shape for which to get predictors}

\item{df}{- whether to return the predictors as data.frame, useful for
predict. If FALSE, returns a named vector}

\item{na_replacement}{- the value to use to replace NAs}
}
\value{
A named vector or dataframe containing all predictors for the shape.
        The list of predictors is contained in ssoftveR::shape_predictors.
}
\description{
Get predictors for a shape
}
