% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{get_shapes}
\alias{get_shapes}
\title{Get a list of shapes on the image}
\usage{
get_shapes(img)
}
\arguments{
\item{img}{- the cimg imager object}
}
\value{
A list of pixsets, each representing one shape (connected set
        of pixels, excluding background)
}
\description{
Gets a list of pixsets representing connected areas in the image, aka shapes
}
