% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_geom.R
\name{get_sides}
\alias{get_sides}
\title{Get sides of shape}
\usage{
get_sides(shape, ma = 7, vel_n = 2)
}
\arguments{
\item{shape}{- the shape for which to find vertices}

\item{ma}{- number of moving average points used for smoothing. Must be odd.}

\item{vel_n}{- number of points before and after current to use for
approximating velocity vector}
}
\value{
A list of objects of class 'side', each containing the x and y
        coordinates of the points of the side of the shape, and also the
        approximate slope angle of the line.
}
\description{
Sides (straight lines in the shape) are sets of points where the slope
doesn't change significantly
}
