% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_geom.R
\name{get_vertices}
\alias{get_vertices}
\title{Get possible vertices}
\usage{
get_vertices(shape, sides = NULL, ...)
}
\arguments{
\item{shape}{- the shape for which to find vertices}

\item{sides}{- the list of sides to use for vertices, if NULL get_sides() is
called to extract sides from shape}

\item{...}{- additional parameters passed to get_lines}
}
\value{
A list containing the x and y coordinates of vertices, and the
        approximate angles at the respective vertices
}
\description{
Possible vertices are points where the slope of the contour of the shape
changes significantly
}
