% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{moving_average}
\alias{moving_average}
\title{Smooth vector using moving averages}
\usage{
moving_average(x, n = 5, cyclic = TRUE)
}
\arguments{
\item{x}{- the vector to smooth}

\item{n}{- the number of points to use for smoothing. Must be odd.}

\item{cyclic}{- logical indicating whether to treat the vector as cyclic}
}
\value{
A vector of centered averages of size n. If cyclic=TRUE, it is the
        same length as vector x and the last elements are averages of first
        and last elements (cyclical overflow).
}
\description{
Smooth vector using moving averages
}
