% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rgb2col}
\alias{rgb2col}
\title{Get the nearest color name from rgb}
\usage{
rgb2col(rgb)
}
\arguments{
\item{rgb}{- vector with r,g,b components, in that order}
}
\description{
Get the nearest color name from rgb
}
