% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{shape_from_pixset}
\alias{shape_from_pixset}
\title{Make a 'shape' object from a pixset}
\usage{
shape_from_pixset(px)
}
\arguments{
\item{px}{- the pixset to turn into a shape}
}
\value{
An object (list) of class 'shape', containing the pixset itself, the
        xy coordinates of the pixset and the contours of the pixset.
}
\description{
Turns a pixset into a object of class 'shape', extracting xy coordinates of
pixels and contours of the pixset
}
