% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_predictors.R
\docType{data}
\name{shape_predictors}
\alias{shape_predictors}
\title{We'll define a vector of names of predictor functions which we use.}
\format{An object of class \code{character} of length 5.}
\usage{
shape_predictors
}
\description{
We'll define a vector of names of predictor functions which we use.
}
\keyword{datasets}
