% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{shift}
\alias{shift}
\title{Shift a vector cyclically}
\usage{
shift(x, n = 0)
}
\arguments{
\item{x}{- the vector to shift}

\item{n}{- the number of places to shift  (can be negative)}
}
\value{
A vector of the same size as x, shifted cyclically by n places (e.g.
        (1,2,3) shifted by n = 1 is (2,3,1))
}
\description{
Shift a vector cyclically
}
