% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_predictors.R
\name{side_length_variance}
\alias{side_length_variance}
\title{Variance of the lengths of sides (distance between vertices)}
\usage{
side_length_variance(shape)
}
\arguments{
\item{shape}{- the shape for which to calculate the predictor}
}
\value{
The variance of the distances between adjacent vertices
}
\description{
The idea is that regular polygons will have all sides same, i.e. var = 0
}
