% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mines.R
\name{slide_matrix_single}
\alias{slide_matrix_single}
\title{Internal function used to slide a matrix in a single direction.}
\usage{
slide_matrix_single(mat, dir, fill)
}
\description{
If you want to slide a matrix, use slide_matrix
}
