% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{solid_blobs}
\alias{solid_blobs}
\title{Find solid color connected areas in the image}
\usage{
solid_blobs(img)
}
\arguments{
\item{img}{- the cimg imager object}
}
\value{
A pixset where every pixel is TRUE if the neighbourhood is of the
        same color as the pixel
}
\description{
Turns the image object into a logical image (pixset) of blobs of pixels of
the same color
}
\details{
Each pixel maps to TRUE if every pixel in the square 3x3 neighbourhood is
the same color, otherwise it becomes FALSE.

Some erosion will happen because of the square neighbourhood
}
