% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_predictors.R
\name{vertex_count}
\alias{vertex_count}
\title{Number of possible vertices}
\usage{
vertex_count(shape)
}
\arguments{
\item{shape}{- the shape for which to calculate the predictor}
}
\value{
The number of possible vertices found (using the get_vertices()
        function)
}
\description{
Number of possible vertices
}
