% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mines.R
\name{clip}
\alias{clip}
\title{Clip image border}
\usage{
clip(im, amount = 0)
}
\arguments{
\item{im}{- the image to clip}

\item{amount}{- the number of pixels to clip, as numeric, or, if character,
the percentage of dimension of the image (e.g.
clip(im, "5%")).}
}
\value{
The clipped image
}
\description{
Clip image around the border given the amount
}
\details{
'amount' can be a single value or a vector of size 4. If only one value is
given, all sides are clipped with the same amount. If it is a vector with
4 values, the sides are clipped by amount, respectively, in the order
left, right, top, bottom.
}
