% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{get_shapes}
\alias{get_shapes}
\title{Get a list of shapes on the image}
\usage{
get_shapes(img, min_area = 25, ...)
}
\arguments{
\item{img}{- the cimg imager object}

\item{min_area}{- the minimum area (number of pixels) of a connected object
that can be considered a shape}

\item{...}{- additional parameters passed to solid_blobs (stencil/cr_n)}
}
\value{
A list of pixsets, each representing one shape (connected set
        of pixels, excluding background)
}
\description{
Gets a list of pixsets representing connected areas in the image
}
