% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{solid_blobs}
\alias{solid_blobs}
\title{Find solid color connected areas in the image}
\usage{
solid_blobs(img, stencil = NULL, cr_n = max(min(dim(img)[1:2])\%/\%100, 5))
}
\arguments{
\item{img}{- the cimg imager object}

\item{stencil}{- a stencil dataframe which defines the neighbourhood we use.
Must have a dx and dy columns, like for the imager package}

\item{cr_n}{- if set, the stencil becomes a "n x n" cross (+) neighbourhood,
centered at current pixel. Should be an odd number.}
}
\value{
A pixset where every pixel is TRUE if the neighbourhood is of the
        same color as the pixel
}
\description{
Turns the image object into a logical image (pixset) of blobs of pixels of
the same color
}
\details{
Each pixel maps to TRUE if every pixel in the stencil neighbourhood is
the same color, otherwise it becomes FALSE.

Some erosion will happen because of the square neighbourhood
}
